--- app/src/main/java/com/marlena/pictures_context_project/ui/scenes/galerymovie/Movie.kt
@@ -1,6 +1,0 @@
-package com.marlena.pictures_context_project.ui.scenes.galerymovie
-
-interface Movie {
-    interface View{}
-    interface Presenter{}
-}
