--- app/src/main/java/com/marlena/pictures_context_project/ui/scenes/galerymovie/MoviePresenter.kt
@@ -1,4 +1,0 @@
-package com.marlena.pictures_context_project.ui.scenes.galerymovie
-
-class MoviePresenter: Movie.Presenter {
-}
