+++ app/src/main/java/com/marlena/pictures_context_project/ui/scenes/gallerycat/Cat.kt
@@ -1,8 +1,8 @@
-package com.marlena.pictures_context_project.ui.scenes.galeryanimal
+package com.marlena.pictures_context_project.ui.scenes.gallerycat
 
 import com.marlena.pictures_context_project.ui.model.ThePicture
 
-interface Animal {
+interface Cat {
 
     interface View{
         fun setAllList(list: List<ThePicture>)
