+++ app/src/main/java/com/marlena/pictures_context_project/ui/scenes/gallerycat/CatPresenter.kt
@@ -1,15 +1,13 @@
-package com.marlena.pictures_context_project.ui.scenes.galeryanimal
+package com.marlena.pictures_context_project.ui.scenes.gallerycat
 
 import com.marlena.pictures_context_project.ui.core.App
-import com.marlena.pictures_context_project.ui.model.PictureResponse
-import com.marlena.pictures_context_project.ui.model.ThePicture
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import kotlin.coroutines.CoroutineContext
 
-class AnimalPresenter(private val view: Animal.View): Animal.Presenter, CoroutineScope {
+class CatPresenter(private val view: Cat.View): Cat.Presenter, CoroutineScope {
     override val coroutineContext: CoroutineContext
         get() = Dispatchers.Main
