+++ app/src/main/java/com/marlena/pictures_context_project/ui/scenes/gallerylandscape/LandscapePresenter.kt
@@ -1,37 +1,18 @@
-package com.marlena.pictures_context_project.ui.scenes.galerymovie
+package com.marlena.pictures_context_project.ui.scenes.gallerylandscape
 
-import android.content.Intent
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.marlena.pictures_context_project.R
 import com.marlena.pictures_context_project.ui.model.ThePicture
-import com.marlena.pictures_context_project.ui.scenes.pictureadapter.PictureAdapter
-import com.marlena.pictures_context_project.ui.scenes.showPicture.PictureActivity
-import kotlinx.android.synthetic.main.fragment_gallery_landscape.*
 
-class MovieFragment : Fragment(), Movie.View, PictureAdapter.Listener {
+class LandscapePresenter : Landscape.Presenter {
 
-    private val actionList: MutableList<ThePicture> by lazy { mutableListOf<ThePicture>() }
+    lateinit var pictureList: ArrayList<ThePicture>
 
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.fragment_gallery_movie, container, false)
-    }
+    override fun getAllList(): ArrayList<ThePicture> {
 
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        val list = arrayListOf(
-//            ThePicture(
-//                url = Constants.imageUrl + movie.posterPath,
-//                name = "Movie"
-//            ),
+        pictureList = arrayListOf(
+            ThePicture(
+                url = "https://image-store.slidesharecdn.com/6d8bec77-43fa-4a9f-b183-e70fa8197397-original.jpeg",
+                name = "FurtaCor"
+            ),
             ThePicture(
                 url = "https://www.photoblog.com/learn/wp-content/uploads/2018/10/mark-harpur-748500-unsplash.jpg",
                 name = "Paz Roxa"
@@ -109,19 +90,6 @@ class MovieFragment : Fragment(), Movie.View, PictureAdapter.Listener {
                 name = "Submarine"
             )
         )
-
-        val adapter = PictureAdapter(
-            list,
-            this
-        )
-        recyclerViewRV?.adapter = adapter
-    }
-
-    override fun openPictureFragment(url: String, name: String) {
-        //transformar em activity, abrir com transparencia
-        val intent = Intent(context, PictureActivity::class.java)
-        intent.putExtra("imageUrl", url)
-        intent.putExtra("imageName", name)
-        startActivity(intent)
+        return pictureList
     }
 }
