+++ app/src/main/java/com/marlena/pictures_context_project/ui/scenes/gallerymovie/Movie.kt
@@ -1,0 +1,6 @@
+package com.marlena.pictures_context_project.ui.scenes.gallerymovie
+
+interface Movie {
+    interface View{}
+    interface Presenter{}
+}
