+++ app/src/main/java/com/marlena/pictures_context_project/ui/scenes/gallerymovie/MoviePresenter.kt
@@ -1,0 +1,4 @@
+package com.marlena.pictures_context_project.ui.scenes.gallerymovie
+
+class MoviePresenter: Movie.Presenter {
+}
