+++ app/src/main/res/menu/activity_main_drawer.xml
@@ -14,7 +14,7 @@
                 <item
                     android:id="@+id/nav_gallery_animal"
                     android:icon="@drawable/ic_menu_gallery"
-                    android:title="@string/menu_gallery_animal" />
+                    android:title="@string/menu_gallery_cat" />
 
                 <item
                     android:id="@+id/nav_gallery_landscape"
