+++ app/src/main/res/values/strings.xml
@@ -8,13 +8,18 @@
     <string name="action_settings">Settings</string>
 
     <string name="menu_home">My Diary</string>
-    <string name="menu_gallery_animal">Animals</string>
+    <string name="menu_gallery_cat">Cats</string>
     <string name="menu_gallery_landscape">Landscape</string>
     <string name="menu_gallery_movie">Movie</string>
     <string name="menu_about">About</string>
 
     <string name="desc_app">App para suas melhores imagens.</string>
-    <string name="welcome_text">Olá!</string>
+    <string name="welcome_text">
+        Expressar o que sente é necessário para a vida de qualquer ser.
+        Respeitar isso e tornar um hábito saudável pode mudá-lo para melhor!
+        Inspire-se com animais, pessoas, músicas, coisas.
+        Use esse aplicativo como forma de demonstrar seus sentimentos e sensações.
+        Anote tudo que sentir com carinho e faça disso um hábito!</string>
     <string name="image_title">Título da Imagem</string>
     <string name="image_id">ID</string>
     <string name="communicate">More</string>
